+++ modules/client/src/addressList/internal/graphql-queries/members.ts
@@ -1,0 +1,11 @@
+import { gql } from "graphql-request";
+
+export const QueryAddressListMembers = gql`
+query AddressListPluginMembers($address: ID!) {
+    addresslistPlugin(id: $address){
+        members {
+            address
+        }
+    }
+}
+`;