+++ modules/client/test/integration/client-erc20/methods.test.ts
@@ -30,7 +30,6 @@ import { InvalidAddressOrEnsError } from "@aragon/sdk-common";
 import {
   contextParams,
   contextParamsLocalChain,
-  TEST_DAO_ADDRESS,
   TEST_ERC20_DAO_ADDRESS,
   TEST_ERC20_PLUGIN_ADDRESS,
   TEST_ERC20_PROPOSAL_ID,
@@ -214,8 +213,8 @@ describe("Client ERC20", () => {
         const ctxPlugin = ContextPlugin.fromContext(ctx);
         const client = new ClientErc20(ctxPlugin);
 
-        const daoAddress = TEST_DAO_ADDRESS;
-        const wallets = await client.methods.getMembers(daoAddress);
+        const pluginAddress = TEST_ERC20_PLUGIN_ADDRESS;
+        const wallets = await client.methods.getMembers(pluginAddress);
 
         expect(Array.isArray(wallets)).toBe(true);
         expect(wallets.length).toBeGreaterThan(0);